requirements.txt
openai-agents
python-dotenv
rich
requests
bs4
pydantic
duckduckgo-search
